chore:SP-2968 Organizes inspection module by data sources
